requirements.txt
kafka-python==2.2.15
python-dotenv==1.1.1
SQLAlchemy==2.0.43
pandas==2.3.2
psycopg2-binary==2.9.10
requests==2.32.5